fix: resolve all cargo clippy warnings
- Add #[allow(clippy::type_complexity)] to structs with complex callback types
- Collapse nested if statements into single if-let chains
- Replace .filter().last()/.next_back() with .rfind()
- Remove unnecessary i64 cast in format.rs
- Collapse consecutive str::replace calls into single replace with array
- Replace assert_eq! with literal bool to assert!/assert!
- Fix unnecessary unwrap after is_some check
- Add #[allow(clippy::wrong_self_convention)] for builder-pattern is_* methods
- Remove if/else with identical branches in note_card.rs

+++ apps/desktop/src/store/format.rs
@@ -2,7 +2,7 @@
 /// ("Just now", "5m ago", "3h ago", "Yesterday", "4d ago").
 pub fn format_relative_time(epoch_ms: i64) -> String {
     let now = current_time_ms();
-    let delta_secs = ((now - epoch_ms).max(0) / 1000) as i64;
+    let delta_secs = (now - epoch_ms).max(0) / 1000;
 
     if delta_secs < 60 {
         return "Just now".to_string();
